+++ CHANGES.rst
@@ -14,6 +14,9 @@ Changes since v1.1
   variable).
 - Provide better error messages when trying to allocate more than 10000 records of a
   forcing field.
+- PISM supports CMake 3.1 again (v1.1 required CMake 3.13 for no good reason).
+- Fix PISM's `-regional` runs: disable ice flow, surface mass balance, and basal mass
+  balance effects on ice geometry in "no model" areas.
 
 Changes from v1.0 to v1.1
 =========================